+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Change Log
 
 ## Version 0.9.17
 
+- Added support for toggling default or hot interests in subject
+  details ([#22](https://github.com/Bumblebee202111/doubean-public/issues/22))
+- Updated interest status logic to match official behavior based on release state in subject details
+- Fixed subject reviews sheet being initially fully
+  expanded ([#22](https://github.com/Bumblebee202111/doubean-public/issues/22))
+
 ## Version 0.9.16
 
 - ([#21](https://github.com/Bumblebee202111/doubean-public/issues/21)) Optimized jumping to comments